Bump item-renderer dependency version. Remove classes from iris that were moved to item-renderer. Update package references to new package names. Add 'tds.blackbox.web' package-scan to iris to ensure moved controller is still registered with Spring.
