--- iris/src/main/java/tds/iris/abstractions/repository/IContentHelper.java
@@ -1,19 +1,0 @@
-/*******************************************************************************
- * Educational Online Test Delivery System 
- * Copyright (c) 2014 American Institutes for Research
- *       
- * Distributed under the AIR Open Source License, Version 1.0 
- * See accompanying file AIR-License-1_0.txt or at
- * http://www.smarterapp.org/documents/American_Institutes_for_Research_Open_Source_Software_License.pdf
- ******************************************************************************/
-package tds.iris.abstractions.repository;
-
-import tds.iris.web.data.ContentRequest;
-import tds.itemrenderer.data.ItemRenderGroup;
-
-public interface IContentHelper
-{
-  public ItemRenderGroup loadRenderGroup (ContentRequest contentRequest);
-
-  public boolean reloadContent ();
-}
