+++ iris/src/main/resources/root-context.xml
@@ -47,6 +47,7 @@
 	<context:component-scan base-package="tds.student.web" />
 	<context:component-scan base-package="tds.student.configuration" />
 	<context:component-scan base-package="tds.itemscoringengine" />
+	<context:component-scan base-package="tds.blackbox.web" />
 
 	<beans:bean id="iTDSLogger" class="tds.iris.logger.IrisLogger" scope="singleton" />
 	<beans:bean id="iContentBuilder" class="tds.iris.content.ContentBuilder" scope="singleton" />
