+++ student/pom.xml
@@ -54,11 +54,6 @@
 
 		<dependency>
 			<groupId>org.opentestsystem.delivery</groupId>
-			<artifactId>student-web</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.opentestsystem.delivery</groupId>
 			<artifactId>item-renderer</artifactId>
 		</dependency>
 
@@ -242,8 +237,8 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.opentestsystem.blackbox</groupId>
-			<artifactId>shared-blackbox</artifactId>
+			<groupId>org.opentestsystem.delivery</groupId>
+			<artifactId>item-renderer-web</artifactId>
 			<type>war</type>
 		</dependency>
