+++ student/src/main/java/tds/student/web/handlers/ContentHandler.java
@@ -8,10 +8,7 @@
  ******************************************************************************/
 package tds.student.web.handlers;
 
-import java.io.IOException;
-
-import javax.servlet.http.HttpServletResponse;
-
+import TDS.Shared.Exceptions.ReturnStatusException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,7 +18,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-
+import tds.blackbox.web.handlers.BaseContentRendererController;
 import tds.itemrenderer.data.AccLookup;
 import tds.itemrenderer.data.AccProperties;
 import tds.itemrenderer.data.IITSDocument;
@@ -30,7 +27,6 @@ import tds.itemrenderer.data.ITSTypes.ITSEntityType;
 import tds.itemrenderer.data.InvalidDataException;
 import tds.itemrenderer.data.ItemRender;
 import tds.itemrenderer.data.ItemRenderGroup;
-import tds.student.services.PrintServiceImpl;
 import tds.student.services.abstractions.IContentService;
 import tds.student.services.abstractions.IResponseService;
 import tds.student.services.abstractions.PrintService;
@@ -41,7 +37,8 @@ import tds.student.web.StudentContext;
 import tds.student.web.StudentContextException;
 import tds.student.web.StudentSettings;
 
-import TDS.Shared.Exceptions.ReturnStatusException;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
 @Scope ("prototype")
 @Controller
