+++ student/src/main/java/tds/student/web/handlers/GlobalHandler.java
@@ -8,12 +8,13 @@
  ******************************************************************************/
 package tds.student.web.handlers;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.UUID;
-
-import javax.servlet.http.HttpServletRequest;
-
+import AIR.Common.TDSLogger.ITDSLogger;
+import AIR.Common.Web.WebHelper;
+import TDS.Shared.Exceptions.ReturnStatusException;
+import TDS.Shared.Messages.IMessageService;
+import TDS.Shared.Messages.MessageJson;
+import TDS.Shared.Messages.MessageSystem;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,20 +22,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-
-import tds.student.services.abstractions.IGlobalService;
+import tds.blackbox.web.handlers.TDSHandler;
 import tds.student.web.controls.dummy.GlobalJavascript;
 import tds.student.web.data.JavascriptStack;
 import tds.student.web.data.JavascriptStackFrame;
-import AIR.Common.TDSLogger.ITDSLogger;
-import AIR.Common.Web.WebHelper;
-import TDS.Shared.Exceptions.ReturnStatusException;
-import TDS.Shared.Messages.IMessageService;
-import TDS.Shared.Messages.MessageJson;
-import TDS.Shared.Messages.MessageSystem;
-import TDS.Shared.Security.TDSIdentity;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+import java.util.List;
 
 /**
  * @author mpatel
